Finalizado Rotas Rest
Concluido o desenvolvimento das funcionalidades via rest, nestes commits
e possivel realizar todos os requisitos funcionais exgidos com escessao
da devolucao de um item especifico, este ponto acabei nao desenvolvendo
pelo tempo.

+++ src/test/java/com/invillia/acme/common/TestCase.java
@@ -32,7 +32,7 @@ public class TestCase {
 			OrderItem item = OrderItem.builder()
 									  .description("Item numero "+i)
 									  .quantity((float) 1)
-									  .unit(2)
+									  .unit((float)2)
 									  .build();
 			items.add(item);
 		}
